+++ examples/injection.py
@@ -78,22 +78,30 @@ async def _fetch_info(database: protos.DatabaseProto = tanjun.injected(type=prot
     raise NotImplementedError  # This is an example callback and doesn't provide an implementation.
 
 
+# Since this is being used as an injected callback, we can also ask for an injected type here.
+async def _fetch_cachable_info(
+    database: protos.DatabaseProto = tanjun.injected(type=protos.DatabaseProto),
+) -> typing.Any:
+    raise NotImplementedError  # This is an example callback and doesn't provide an implementation.
+
+
 @component.with_command
 @tanjun.as_slash_command("info", "Command description")
 async def get_info(
     ctx: tanjun.abc.MessageContext,
-    # Here we set _fetch_info as a cached injected callback.
+    # Here we set _fetch_info as an injected callback.
     #
     # Injected callbacks are callbacks which'll be called before this function is called
     # with the result of it being being passed to the command callback.
     # (note these also support dependency injection).
+    info: typing.Any = tanjun.injected(callback=_fetch_info),
+    # Here we set _fetch_cachable_info as a cached injected callback.
     #
-    # cache_callback is used to cache the result of _fetch_info for 30 minutes, meaning
-    # that it'll be called at most once every 30 minutes for this command.
-    info: typing.Any = tanjun.injected(
-        callback=tanjun.cache_callback(_fetch_info, expire_after=datetime.timedelta(minutes=30))
-    ),
+    # `cached_inject(callback)` is a variant of `inject(callback=callback)`
+    # which caches the result of the callback for the expire duration.
+    cached_info: typing.Any = tanjun.cached_inject(_fetch_cachable_info, expire_after=datetime.timedelta(minutes=30)),
 ) -> None:
+    await ctx.respond(cached_info.format())
     await ctx.respond(info.format())