+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `tanjun.dependencies.injected_lc` which is a shortcut for
+- `tanjun.dependencies.inject_lc(Type)` which is a shorthand for
   `tanjun.injected(callback=tanjun.make_lc_resolver(Type))`.
+- `tanjun.dependencies.cached_inject(...)` which is a shorthand for
+  `tanjun.injected(callback=tanjun.cache_callback(...))
 
 ## [2.1.2a1] - 2021-10-15
 ### Added